package.bbclass: Move from BB_TASKHASH to BB_UNIHASH for PR service
An unusual corner case can result in a non-deterministic result.  The case
can be reproduced by doing the following:

enable PR Service, and hash equivalency then run

bitbake glibc

edit the meta/recipes-core/glibc/glibc_2.32.bb, add a comment to
the do_patch_append task.

Run the following again

bitbake glibc

If you inspect the log, the system should show a unihash was determined in
the do_package task.

Now to trigger the corner case, KEEP your 'cache' directory, so the PR
database and unihash database remains, but remove you sstate-cache directory.

If you run a build before this change, the PR value for the glibc will be r0.1,
instead of r0.0 as expected (based on the original build.)

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>